requirement.txt
jsonrpclib-pelix
pymongo
pika
requests
redis
lxml
sklearn
python-dateutil
newspaper3k
dateutil
